[FileTree.vue] Do not delete deletable if not in persistent store.

+++ RefMan/Client/src/components/references/file-system/FileTree.vue
@@ -103,7 +103,9 @@ export default {
       const currentNodeParent = currentNode.parent;
       currentNodeParent.remove(currentNode);
 
-      await focal.deletable.delete();
+      if (currentNode.existsInPersistentStore) {
+        await focal.deletable.delete();
+      }
     },
     addNewNode(node, addFunction) {
       const closestFolderToFocus = this.findClosestFolderToFocus();
